Support showOutlierRange parameter. Update unit tests.
